requirements.txt
SQLAlchemy==2.0.23
python-dotenv==1.0.0
Flask==3.0.0
rapidfuzz==3.5.2
inflect==7.0.0
ollama==0.3.0
spacy==3.7.2
pyspellchecker==0.8.1
urllib3<2.0.0
PyYAML==6.0.1